Add views for schedules
The /schedules view shows all schedules configured in the DB; this
is a reasonable default until it needs to be locked down.

The /schedules/<id> view shows a "table" of dates to classes. The
logic currently skips Saturdays and any date included in a pause
window. The logic is not yet checking for holidays or coallescing
multiple part classes into one row entry.

The end goal of all this is to be able to create an RSS feed for
each schedule with the list of classes it should have. Light
refactoring will be needed to get there.